fix(types): resolve all 30 mypy errors surfaced by full type-check
make verify's mypy step is advisory (tail -1), so these errors were
never surfacing. Running the full `mypy rheojax` found 30 real errors
across 13 files, mostly None-narrowing gaps around get_value()
(float | None) and a few list/tuple vs ndarray indexing false
positives. Fixed the root cause in each case rather than suppressing:

- BaseModel._standard_nlsq_fit now returns Self instead of BaseModel,
  fixing 5 subclass return-type errors in one place (flow models)
- RheoData.y_real/y_imag: move the existing runtime type-narrowing
  check earlier so it also covers the (N,2) fast path
- Registry.__abstractmethods__ access via getattr (not in typeshed)
- SRFS._raise_missing_sgr_params annotated NoReturn so callers narrow
  x/tau0 correctly after the None guard
- Remaining fixes are None-guards or explicit casts at the actual
  point where an Optional/union value meets a strict numeric API

No behavior changes; make verify (lint + mypy + smoke tests, 2037
passed) is green.

+++ rheojax/models/dmt/nonlocal_model.py
@@ -511,7 +511,7 @@ class DMTNonlocal(DMTBase):
             band_location = self.y[band_indices].mean()
         else:
             band_width = self.gap_width
-            band_location = self.gap_width / 2
+            band_location = np.float64(self.gap_width / 2)
 
         return {
             "is_banding": is_banding,
